Proper Agathion slot mask.

+++ L2J_Mobius_5.0_Salvation/java/com/l2jmobius/gameserver/data/xml/impl/AppearanceItemData.java
@@ -102,7 +102,7 @@ public class AppearanceItemData implements IGameXmlReader
 								}
 								case "bodyPart":
 								{
-									final int part = ItemTable.SLOTS.get(c.getTextContent());
+									final long part = ItemTable.SLOTS.get(c.getTextContent());
 									stone.addBodyPart(part);
 									break;
 								}
