+++ L2J_Mobius_5.0_Salvation/java/com/l2jmobius/gameserver/model/actor/instance/L2PcInstance.java
@@ -2197,7 +2197,7 @@ public final class L2PcInstance extends L2Playable
 			}
 			sendPacket(sm);
 			
-			final int slot = _inventory.getSlotFromItem(item);
+			final long slot = _inventory.getSlotFromItem(item);
 			// we can't unequip talisman by body slot
 			if ((slot == L2Item.SLOT_DECO) || (slot == L2Item.SLOT_BROOCH_JEWEL) || (slot == L2Item.SLOT_AGATHION))
 			{
@@ -4990,7 +4990,7 @@ public final class L2PcInstance extends L2Playable
 				}
 				else
 				{
-					final int slot = _inventory.getSlotFromItem(_inventory.getItemByItemId(9819));
+					final long slot = _inventory.getSlotFromItem(_inventory.getItemByItemId(9819));
 					_inventory.unEquipItemInBodySlot(slot);
 					destroyItem("CombatFlag", _inventory.getItemByItemId(9819), null, true);
 				}
@@ -10887,7 +10887,7 @@ public final class L2PcInstance extends L2Playable
 				}
 				else
 				{
-					final int slot = _inventory.getSlotFromItem(_inventory.getItemByItemId(9819));
+					final long slot = _inventory.getSlotFromItem(_inventory.getItemByItemId(9819));
 					_inventory.unEquipItemInBodySlot(slot);
 					destroyItem("CombatFlag", _inventory.getItemByItemId(9819), null, true);
 				}
